+++ tests/test_annotations.py
@@ -165,7 +165,7 @@ class TestAnnotations:
     @pytest.mark.parametrize("slots", [True, False])
     def test_auto_attribs_subclassing(self, slots):
         """
-        Attributes from superclasses are inherited, it doesn't matter if the
+        Attributes from base classes are inherited, it doesn't matter if the
         subclass has annotations or not.
 
         Ref #291
@@ -251,9 +251,9 @@ class TestAnnotations:
         assert c.x == 0
         assert c.y == 1
 
-    def test_super_class_variable(self):
+    def test_base_class_variable(self):
         """
-        Superclass class variables can be overridden with an attribute
+        Base class' class variables can be overridden with an attribute
         without resorting to using an explicit `attr.ib()`.
         """