+++ tests/test_dark_magic.py
@@ -47,7 +47,7 @@ class C2Slots(object):
 
 
 @attr.s
-class Super(object):
+class Base(object):
     x = attr.ib()
 
     def meth(self):
@@ -55,7 +55,7 @@ class Super(object):
 
 
 @attr.s(slots=True)
-class SuperSlots(object):
+class BaseSlots(object):
     x = attr.ib()
 
     def meth(self):
@@ -63,12 +63,12 @@ class SuperSlots(object):
 
 
 @attr.s
-class Sub(Super):
+class Sub(Base):
     y = attr.ib()
 
 
 @attr.s(slots=True)
-class SubSlots(SuperSlots):
+class SubSlots(BaseSlots):
     y = attr.ib()
 
 
@@ -203,7 +203,7 @@ class TestDarkMagic(object):
     @pytest.mark.parametrize("cls", [Sub, SubSlots])
     def test_subclassing_with_extra_attrs(self, cls):
         """
-        Sub-classing (where the subclass has extra attrs) does what you'd hope
+        Subclassing (where the subclass has extra attrs) does what you'd hope
         for.
         """
         obj = object()
@@ -215,10 +215,10 @@ class TestDarkMagic(object):
         else:
             assert "SubSlots(x={obj}, y=2)".format(obj=obj) == repr(i)
 
-    @pytest.mark.parametrize("base", [Super, SuperSlots])
+    @pytest.mark.parametrize("base", [Base, BaseSlots])
     def test_subclass_without_extra_attrs(self, base):
         """
-        Sub-classing (where the subclass does not have extra attrs) still
+        Subclassing (where the subclass does not have extra attrs) still
         behaves the same as a subclass with extra attrs.
         """
 
@@ -259,8 +259,8 @@ class TestDarkMagic(object):
             C1Slots,
             C2,
             C2Slots,
-            Super,
-            SuperSlots,
+            Base,
+            BaseSlots,
             Sub,
             SubSlots,
             Frozen,
@@ -283,8 +283,8 @@ class TestDarkMagic(object):
             C1Slots,
             C2,
             C2Slots,
-            Super,
-            SuperSlots,
+            Base,
+            BaseSlots,
             Sub,
             SubSlots,
             Frozen,
@@ -342,11 +342,11 @@ class TestDarkMagic(object):
     @pytest.mark.parametrize("weakref_slot", [True, False])
     def test_attrib_overwrite(self, slots, frozen, weakref_slot):
         """
-        Subclasses can overwrite attributes of their superclass.
+        Subclasses can overwrite attributes of their base class.
         """
 
         @attr.s(slots=slots, frozen=frozen, weakref_slot=weakref_slot)
-        class SubOverwrite(Super):
+        class SubOverwrite(Base):
             x = attr.ib(default=attr.Factory(list))
 
         assert SubOverwrite([]) == SubOverwrite()
@@ -419,9 +419,9 @@ class TestDarkMagic(object):
 
         assert hash(C()) != hash(C())
 
-    def test_overwrite_super(self):
+    def test_overwrite_base(self):
         """
-        Superclasses can overwrite each other and the attributes are added
+        Base classes can overwrite each other and the attributes are added
         in the order they are defined.
         """