+++ CHANGELOG.rst
@@ -192,7 +192,7 @@ Changes
   (`#198 <https://github.com/python-attrs/attrs/issues/198>`_)
 - ``attr.Factory`` is hashable again.
   (`#204 <https://github.com/python-attrs/attrs/issues/204>`_)
-- Subclasses now can overwrite attribute definitions of their superclass.
+- Subclasses now can overwrite attribute definitions of their base classes.
 
   That means that you can -- for example -- change the default value for an attribute by redefining it.
   (`#221 <https://github.com/python-attrs/attrs/issues/221>`_, `#229 <https://github.com/python-attrs/attrs/issues/229>`_)